Add planning functionality to workouts
I find it useful to plan a few movements ahead of time before working out
so I can just get into it better when I go. This is a pretty bare-bones
version of that, you just call "plan" and it writes a simple text file
that is just a list of movements. Then when you workout, if you select
that plan it will list those workouts at the top, then move them off
once you do them.